improvement(helm): only fetch charts when needed
We add an async lock at the module level to the `prepareHelmModule`
handler, which prevents possible race conditions e.g. when several test
suites for a given Helm module are run concurrently.

Also added a fetch indicator file to the build directory of Helm modules
using remote charts to avoid fetching the chart multiple times in the
same situation (e.g. when multiple tests suites are run concurrently for
a given Helm module).